Module Opentelemetry_lwt.Metrics

include module type of struct include Opentelemetry.Metrics end

Metrics.

See the spec

type t

A single metric, measuring some time-varying quantity or statistical distribution. It is composed of one or more data points that have precise values and time stamps. Each distinct metric should have a distinct name.

val float : ?start_time_unix_nano:Opentelemetry.Timestamp_ns.t -> ?now:Opentelemetry.Timestamp_ns.t -> ?attrs: (string * [< `Bool of bool | `Int of int | `None | `String of string ]) list -> float -> Opentelemetry__Metrics_types.number_data_point

Number data point, as a float

val int : ?start_time_unix_nano:Opentelemetry.Timestamp_ns.t -> ?now:Opentelemetry.Timestamp_ns.t -> ?attrs: (string * [< `Bool of bool | `Int of int | `None | `String of string ]) list -> int -> Opentelemetry__Metrics_types.number_data_point

Number data point, as an int

val gauge : name:string -> ?description:string -> ?unit_:string -> Opentelemetry__Metrics_types.number_data_point list -> t

Aggregation of a scalar metric, always with the current value

type aggregation_temporality =

| Aggregation_temporality_unspecified

| Aggregation_temporality_delta

| Aggregation_temporality_cumulative

val sum : name:string -> ?description:string -> ?unit_:string -> ?aggregation_temporality:aggregation_temporality -> ?is_monotonic:bool -> Opentelemetry__Metrics_types.number_data_point list -> t

Sum of all reported measurements over a time interval

val histogram_data_point : ?start_time_unix_nano:Opentelemetry.Timestamp_ns.t -> ?now:Opentelemetry.Timestamp_ns.t -> ?attrs: (string * [< `Bool of bool | `Int of int | `None | `String of string ]) list -> ?exemplars:Opentelemetry__Metrics_types.exemplar list -> ?explicit_bounds:float list -> ?sum:float -> bucket_counts:int64 list -> count:int64 -> unit -> Opentelemetry__Metrics_types.histogram_data_point

Histogram data

parameter count
number of values in population (non negative)

parameter sum
sum of values in population (0 if count is 0)

parameter bucket_counts
count value of histogram for each bucket. Sum of the counts must be equal to count. length must be 1+length explicit_bounds

parameter explicit_bounds
strictly increasing list of bounds for the buckets

val histogram : name:string -> ?description:string -> ?unit_:string -> ?aggregation_temporality:Opentelemetry__Metrics_types.aggregation_temporality -> Opentelemetry__Metrics_types.histogram_data_point list -> t

val make_resource_metrics : ?service_name:string -> ?attrs: (string * [< `Bool of bool | `Int of int | `None | `String of string ]) list -> t list -> Opentelemetry__Metrics_types.resource_metrics

Aggregate metrics into a Proto.Metrics.resource_metrics

val emit : ?attrs: (string * [< `Bool of bool | `Int of int | `None | `String of string ]) list -> t list -> unit

Emit some metrics to the collector (sync). This blocks until the backend has pushed the metrics into some internal queue, or discarded them.

NOTE be careful not to call this inside a Gc alarm, as it can cause deadlocks.