Module Opentelemetry_lwt.Sdk

include module type of struct include Opentelemetry.Sdk end

include module type of struct include Opentelemetry.Sdk end

val present : unit -> bool

val get : unit -> Opentelemetry_core.Exporter.t option

val active : unit -> Opentelemetry_util.Aswitch.t

Aswitch of the installed exporter, or Aswitch.dummy if none.

val add_on_tick_callback : (unit -> unit) -> unit

val run_tick_callbacks : unit -> unit

val tick : unit -> unit

Tick all providers and run all registered callbacks. Call this periodically (e.g. every 500ms) to drive metrics collection, GC metrics, and batch timeout flushing. This is the single function client libraries should call from their ticker.

val set : ?traces:Opentelemetry.Provider_config.t -> ?metrics:Opentelemetry.Provider_config.t -> ?logs:Opentelemetry.Provider_config.t -> Opentelemetry_core.Exporter.t -> unit

val self_metrics : unit -> Opentelemetry_core.Metrics.t list

val setup_self_metrics : unit -> unit

Regularly emit metrics about the OTEL SDK. Idempotent.

val get_tracer : ?name:string -> ?version:string -> ?attrs:(string * [< Opentelemetry_core.Value.t ]) list -> ?__MODULE__:string -> unit -> Opentelemetry.Tracer.t

Get a tracer forwarding to the current main exporter.

val get_meter : ?name:string -> ?version:string -> ?attrs:(string * [< Opentelemetry_core.Value.t ]) list -> ?__MODULE__:string -> unit -> Opentelemetry.Meter.t

Get a meter forwarding to the current main exporter.

val get_logger : ?name:string -> ?version:string -> ?attrs:(string * [< Opentelemetry_core.Value.t ]) list -> ?__MODULE__:string -> unit -> Opentelemetry.Logger.t

Get a logger forwarding to the current main exporter.

val self_debug_to_stderr : ?min_level:Opentelemetry.Self_debug.level -> unit -> unit

val remove : unit -> unit Lwt.t